[Performance][MoE] Preserve top-k routing weights without casts (#15299)
## What this PR does

- preserves selector-produced top-k routing weights across the
refactored main-branch router, MoE communication, and quantization paths
- removes intermediate dtype casts from unquantized, W4A4, W4A8, W4A16,
W8A8, MXFP4, and MXFP8 expert paths
- casts routing weights only at operator boundaries whose dtype
contracts require it
- updates deterministic DeepSeek ACLGraph precision baselines
- adds main-specific unit coverage for router and routed-experts dtype
preservation

## Main branch adaptation

The main branch moved quantization methods into per-scheme subpackages
and introduced central routing in `AscendGroupedTopKRouter` and
`AscendRoutedExperts`. This PR maps the release change to those new
paths and removes the two new central intermediate casts as well.

## Validation

- Ruff lint: passed
- Ruff format check: passed
- Python compileall: passed
- Full pre-commit and NPU matrix: delegated to PR CI

Related release PR: #15234


- vLLM main:
https://github.com/vllm-project/vllm/commit/ba07e4a48fc951300d97eb506217dd530583dea3

+++ vllm_ascend/ops/fused_moe/token_dispatcher.py
@@ -425,7 +425,7 @@ class TokenDispatcherWithAllGather(MoETokenDispatcher[MoEAllGatherCombineMetadat
         final_hidden_states = DeviceOperator.npu_moe_token_unpermute(
             permuted_tokens=hidden_states,
             sorted_indices=combine_metadata.expanded_row_idx,
-            probs=combine_metadata.topk_weights,
+            probs=combine_metadata.topk_weights.to(hidden_states.dtype),
         )
         if len(combine_metadata.restore_shape) == 3:
             final_hidden_states = final_hidden_states.view(combine_metadata.restore_shape)
@@ -710,7 +710,7 @@ class TokenDispatcherWithAll2AllV(MoETokenDispatcher[MoEAllToAllCombineMetadata]
         output = torch_npu.npu_moe_token_unpermute(
             permuted_tokens=permutated_local_input_tokens,
             sorted_indices=combine_metadata.reversed_local_input_permutation_mapping.to(torch.int32),
-            probs=combine_metadata.topk_weights,
+            probs=combine_metadata.topk_weights.to(permutated_local_input_tokens.dtype),
             restore_shape=combine_metadata.hidden_shape_before_permute,
         )
         output = output.view(combine_metadata.hidden_shape)
